consolidate fake database and utils scripts
